Changes Applied to w30
Applied page refresh after slot booked,,
changed distance filter logic.
Converting first letter as Capital letter in the company name in the
customer details view.
set opacity to the book button when slots are fully booked.

+++ app/src/main/java/com/sms/within30/dataobjects/CustomerDO.java
@@ -22,7 +22,7 @@ public class CustomerDO {
     Integer slotsAvailable = 0;
     private int expectedTime;
     public String subdomain;
-    public int destinationDistance;
+    public Double destinationDistance;
 
 
 
@@ -122,11 +122,11 @@ public class CustomerDO {
         this.expectedTime = expectedTime;
     }
 
-    public int getDestinationDistance() {
+    public double getDestinationDistance() {
         return destinationDistance;
     }
 
-    public void setDestinationDistance(int destinationDistance) {
+    public void setDestinationDistance(double destinationDistance) {
         this.destinationDistance = destinationDistance;
     }
